// shared/models/auth.ts
// Tipos de usuário usados na aplicação
export type User = {
  id: string;
  email: string | null;
  firstName: string | null;
  lastName: string | null;
  profileImageUrl: string | null;
  createdAt?: Date | null;
  updatedAt?: Date | null;
};

export type UpsertUser = Omit<User, "createdAt" | "updatedAt">;